refactor: rename SuperChartKernel to SuperChartCore
