+++ packages/superset-ui-chart/test/components/SuperChartCore.test.tsx
@@ -7,10 +7,10 @@ import {
   LazyChartPlugin,
   SlowChartPlugin,
 } from './MockChartPlugins';
-import SuperChartKernel from '../../src/components/SuperChartKernel';
+import SuperChartCore from '../../src/components/SuperChartCore';
 import promiseTimeout from './promiseTimeout';
 
-describe('SuperChartKernel', () => {
+describe('SuperChartCore', () => {
   const chartProps = new ChartProps();
 
   const plugins = [
@@ -34,7 +34,7 @@ describe('SuperChartKernel', () => {
   describe('registered charts', () => {
     it('renders registered chart', () => {
       const wrapper = shallow(
-        <SuperChartKernel chartType={ChartKeys.DILIGENT} chartProps={chartProps} />,
+        <SuperChartCore chartType={ChartKeys.DILIGENT} chartProps={chartProps} />,
       );
 
       return promiseTimeout(() => {
@@ -42,7 +42,7 @@ describe('SuperChartKernel', () => {
       });
     });
     it('renders registered chart with lazy loading', () => {
-      const wrapper = shallow(<SuperChartKernel chartType={ChartKeys.LAZY} />);
+      const wrapper = shallow(<SuperChartCore chartType={ChartKeys.LAZY} />);
 
       return promiseTimeout(() => {
         expect(wrapper.render().find('div.test-component')).toHaveLength(1);
@@ -50,14 +50,14 @@ describe('SuperChartKernel', () => {
     });
     it('does not render if chartType is not set', () => {
       // @ts-ignore chartType is required
-      const wrapper = shallow(<SuperChartKernel />);
+      const wrapper = shallow(<SuperChartCore />);
 
       return promiseTimeout(() => {
         expect(wrapper.render().children()).toHaveLength(0);
       }, 5);
     });
     it('adds id to container if specified', () => {
-      const wrapper = shallow(<SuperChartKernel chartType={ChartKeys.DILIGENT} id="the-chart" />);
+      const wrapper = shallow(<SuperChartCore chartType={ChartKeys.DILIGENT} id="the-chart" />);
 
       return promiseTimeout(() => {
         expect(wrapper.render().attr('id')).toEqual('the-chart');
@@ -65,7 +65,7 @@ describe('SuperChartKernel', () => {
     });
     it('adds class to container if specified', () => {
       const wrapper = shallow(
-        <SuperChartKernel chartType={ChartKeys.DILIGENT} className="the-chart" />,
+        <SuperChartCore chartType={ChartKeys.DILIGENT} className="the-chart" />,
       );
 
       return promiseTimeout(() => {
@@ -74,7 +74,7 @@ describe('SuperChartKernel', () => {
     });
     it('uses overrideTransformProps when specified', () => {
       const wrapper = shallow(
-        <SuperChartKernel
+        <SuperChartCore
           chartType={ChartKeys.DILIGENT}
           overrideTransformProps={() => ({ message: 'hulk' })}
         />,
@@ -94,7 +94,7 @@ describe('SuperChartKernel', () => {
         payload: { message: 'hulk' },
       });
       const wrapper = shallow(
-        <SuperChartKernel
+        <SuperChartCore
           chartType={ChartKeys.DILIGENT}
           preTransformProps={() => chartPropsWithPayload}
           overrideTransformProps={props => props.payload}
@@ -112,7 +112,7 @@ describe('SuperChartKernel', () => {
     });
     it('uses postTransformProps when specified', () => {
       const wrapper = shallow(
-        <SuperChartKernel
+        <SuperChartCore
           chartType={ChartKeys.DILIGENT}
           postTransformProps={() => ({ message: 'hulk' })}
         />,
@@ -128,30 +128,28 @@ describe('SuperChartKernel', () => {
       });
     });
     it('renders if chartProps is not specified', () => {
-      const wrapper = shallow(<SuperChartKernel chartType={ChartKeys.DILIGENT} />);
+      const wrapper = shallow(<SuperChartCore chartType={ChartKeys.DILIGENT} />);
 
       return promiseTimeout(() => {
         expect(wrapper.render().find('div.test-component')).toHaveLength(1);
       });
     });
     it('does not render anything while waiting for Chart code to load', () => {
-      const wrapper = shallow(<SuperChartKernel chartType={ChartKeys.SLOW} />);
+      const wrapper = shallow(<SuperChartCore chartType={ChartKeys.SLOW} />);
 
       return promiseTimeout(() => {
         expect(wrapper.render().children()).toHaveLength(0);
       });
     });
     it('eventually renders after Chart is loaded', () => {
-      const wrapper = shallow(<SuperChartKernel chartType={ChartKeys.SLOW} />);
+      const wrapper = shallow(<SuperChartCore chartType={ChartKeys.SLOW} />);
 
       return promiseTimeout(() => {
         expect(wrapper.render().find('div.test-component')).toHaveLength(1);
       }, 1500);
     });
     it('does not render if chartProps is null', () => {
-      const wrapper = shallow(
-        <SuperChartKernel chartType={ChartKeys.DILIGENT} chartProps={null} />,
-      );
+      const wrapper = shallow(<SuperChartCore chartType={ChartKeys.DILIGENT} chartProps={null} />);
 
       return promiseTimeout(() => {
         expect(wrapper.render().find('div.test-component')).toHaveLength(0);
@@ -161,7 +159,7 @@ describe('SuperChartKernel', () => {
 
   describe('unregistered charts', () => {
     it('renders error message', () => {
-      const wrapper = mount(<SuperChartKernel chartType="4d-pie-chart" chartProps={chartProps} />);
+      const wrapper = mount(<SuperChartCore chartType="4d-pie-chart" chartProps={chartProps} />);
 
       return promiseTimeout(() => {
         expect(wrapper.render().find('.alert')).toHaveLength(1);
@@ -171,7 +169,7 @@ describe('SuperChartKernel', () => {
 
   describe('.processChartProps()', () => {
     it('use identity functions for unspecified transforms', () => {
-      const chart = new SuperChartKernel({
+      const chart = new SuperChartCore({
         chartType: ChartKeys.DILIGENT,
       });
       const chartProps2 = new ChartProps();
